backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
redis==5.0.1
celery==5.3.4
flower==2.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
requests==2.31.0
python-dotenv==1.0.0
fastapi-mail==1.4.1
paho-mqtt==2.0.0
openpyxl==3.1.2
numpy==1.24.3
pandas==2.0.3
chardet==5.2.0
